core/ delete header core directory, move content into netero's include directory, delete internal.hpp
