+++ include/netero/netero.hpp
@@ -1,6 +1,6 @@
 /**
  * Netero sources under BSD-3-Clause
- * see LICENCE.txt
+ * see LICENSE.txt
  */
 
 #pragma once
@@ -10,11 +10,9 @@
  * @brief Include all netero core features.
  */
 
-#include <netero/core/internal.hpp>
-#include <netero/core/type_id.hpp>
-
+#include <netero/type_id.hpp>
 #include <netero/observer/signal.hpp>
-#include <netero/observer/slot.hpp>
+#include <netero/observer/signal.hpp>
 
 /**
  * @brief Netero namespace.
@@ -22,7 +20,11 @@
 namespace netero {
 	constexpr char const *version = "0.0.1"; /**< Version string. */
 
+	/**
+	 * @brief Constant numbers namespace.
+	 */
 	namespace numbers {
-        constexpr float pi = 3.14159265358979323846;
+        constexpr float pi = 3.14159265358979323846; /**< PI as float. */
 	}
 }
+
