+++ modules/core/CMakeLists.txt
@@ -12,8 +12,7 @@ message(STATUS "Configure Netero libs")
 
 list(APPEND PUBLIC_HEADER
         ${HEADERS_DIRECTORY}/netero.hpp
-        ${HEADERS_DIRECTORY}/core/internal.hpp
-        ${HEADERS_DIRECTORY}/core/type_id.hpp)
+        ${HEADERS_DIRECTORY}/type_id.hpp)
 
 ##====================================
 ##  Algorithms Sources
